Add optional remote-host mapping
When using gerrit, the git remote url looks something like:
 "ssh://username@gerrit.example.com:29418/project-name"

It is common to have gerrit replicate to github or gitlab, this change
makes it possible to add this mapping from gerrit host to another such
service.  For example, mapping the above url to:
 "https://gitlab.example.com/org-name/project-name"

Changed the url regex to use url-parse, since the port was being
included as part of the path.