Get the current file commit link if pt not on hash
- This functionality if off by default for backward compatibility.
- Set the new variable git-link-commit-fallback-use-latest-commit to t
  to enable this.

Personally I have never needed to get commit link for a hash under
point. But I do find it useful to jump to the commit details associated
to the commit of the current file in buffer (especially when studying
emacs sources).

(Next PR is making git-link work with the emacs git source:
git.savannah.gnu.org)