Allow subdomain git remotes
The `git-link` and `git-commit-link` functions will now work
even if the remote-host is not exactly "gitlab.com", but
something like "gitlab.COMPANY.com" or "gitlab.COMPANY.org" or
"gitlab.COMPANY.co.in".

Updated README.md to reflect these changes.

This commits adds the use of `cl-find-if` macro. If using emacs 24.x or
newer, this is in-built in emacs in `cl-seq` library (which is loaded
by the cl-lib library).

But for older emacs versions, the `cl-lib` library will need to be
auto-installed from GNU Elpa with the help of the Package-Requires
line added in this commit:

    ;; Package-Requires: ((cl-lib "0.6.1"))