Use commit when branch is empty
Pass nil to the link builder if the branch could not be determined.

git-link--branch returns nil when it cannot not determine the current branch.
However, url-hexify-string was converting nil to the empty string.
Consequently, some link handlers were using the empty string as the branch name
instead of falling back to the commit.