Add support for Magit-Blob buffers
Use the correct branch or commit when the current buffer is visiting a blob
using Magit.

* git-link.el (git-link--commit): Use the magit-buffer-revision variable if
the buffer is using Magit-Blob mode.
(git-link--branch): Look up the branch using magit-buffer-revision and the
magit-rev-branch function if the buffer is using Magit-Blob mode.
(git-link--relative-filename): Use the magit-buffer-file-name variable the
buffer is using Magit-Blob mode.
(git-link--using-magit-blob-mode): New function.
(git-link): Set the region and ignore the branch that is checked out into
the working tree when using Magit-Blob mode.