Create a function to return the git-link itself
`git-link` is incredibly useful, but it doesn't return the URL it
creates. This introduces a new function which does so, allowing more
tooling to make use of this great feature.

Wrapped up in this I did some refactoring to drop the use of `setq`
internally to the function, as that defies the purely functional
nature of the call. This _shouldn't_ break anything.