Magit and Dired: fall back to current directory if not on file
This is most noticeable in magit-status, where you'd have to visit a
random file to be able to jump to the repository page with git-link.
In Dired, it's also noticeable if git-link is run when point isn't
over a file, for instance on Dired's current directory heading.

This commit makes it so that, for example, running git-link in
magit-status will fall back to visiting the repository at the current
branch or commit. Then, as a special case, if current directory is
project root (ie. `git-link--relative-filename` returns ""), visit the
homepage instead.

Now git-link in Magit and Dired always goes somewhere sensible,
instead of returning an error.